Added Helmet Package, Extrapolated Logic from App into Main, added Helmet to App, configured Meta Tags for SEO
