+++ package.json
@@ -17,6 +17,7 @@
     "react": "^18.2.0",
     "react-chartjs-2": "^2.11.1",
     "react-dom": "^18.2.0",
+    "react-helmet": "^6.1.0",
     "react-icons": "^4.7.1",
     "react-redux": "^8.0.5",
     "react-scripts": "5.0.1",
