+++ src/components/UserOptions/UserOptions.jsx
@@ -1,12 +1,10 @@
 import { ThemeContext } from '../../contexts/ThemeContext';
 import { useContext } from 'react';
-import TokenDropdown from '../Dropdown/TokenDropdown';
-import ChartDropdown from '../Dropdown/ChartDropdown';
+import TokenDropdown from '../Dropdowns/TokenDropdown';
+import ChartDropdown from '../Dropdowns/ChartDropdown';
 
 function UserOptions({
-  pair,
   handleSelect,
-  currencies,
   granularity,
   handleChart,
   activeGranularity,
@@ -25,7 +23,7 @@ function UserOptions({
     <div className="grid grid-cols-2 grid-rows-2 gap-2.5 w-auto h-auto ">
       <div className="row-start-1 row-span-1 col-start-1 col-end-3 text-xs md:text-lg">
         <h1
-          className={`border-b-4  ${
+          className={`border-b-4 ${
             isDarkMode ? 'border-gray-800' : 'text-sky-300 border-gray-300'
           }`}
         >
@@ -43,58 +41,18 @@ function UserOptions({
         <TokenDropdown
           className="z-10 col-start-1 col-end-1 lg:col-start-2 lg:col-end-2 row-start-2 row-span-2"
           handleSelect={handleSelect}
-          currencies={currencies}
-          activePair={pair}
         />
       </div>
-
-      {/* <select
-        id="cur-select"
-        className={`cursor-pointer md:pl-2 text-gray-800 text-xs rounded border border-orange-500 focus:outline-none ${
-          checkPair(pair) ? 'animate-bounce' : ''
-        }`}
-        name="currency"
-        value={pair}
-        onChange={handleSelect}
-      >
-        <option defaultValue={true}>Select</option>
-        {currencies.map((cur, idx) => {
-          let newIcon = findIcon(cur.base_currency);
-          return (
-            <option key={cur.id} value={cur.id}>
-              {`${newIcon && newIcon}`}
-            </option>
-          );
-        })}
-      </select> */}
-
       <label
         className="row-start-3 py-1 row-span-3 col-start-1 col-end-1 font-bold text-sm hidden lg:text-md lg:block"
         htmlFor="chart-select"
       >
         Chart Granularity
       </label>
-      {/* <select
-        id="chart-select"
-        className="py-1.5 cursor-pointer md:pl-2 text-gray-800 text-xs border-none rounded border focus:outline-none"
-        name="chart"
-        // value={value}
-        onChange={handleChart}
-      >
-        <option defaultValue={true}>Granularity</option>
-        {granularity.map((gran, idx) => {
-          return (
-            <option key={idx} value={gran[0]}>
-              {gran[1]}
-            </option>
-          );
-        })}
-      </select> */}
       <div className="row-start-3 row-span-3 col-start-1 col-end-1 lg:row-start-3 lg:row-span-3 lg:col-start-2 lg:col-end-2">
         <ChartDropdown
           handleChart={handleChart}
           granularity={granularity}
-          activePair={pair}
           activeGranularity={activeGranularity}
         />
       </div>
